pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-science-exam"
description = "Package of llm-science-exam"
readme = "README.md"
requires-python = ">= 3.10"
version = "0.0.1"
dependencies = [
    "faiss-gpu==1.7.2",
    "datasets==2.14.4",
]

[project.optional-dependencies]
dev = [
    "tomli==2.0.1",
    "tqdm==4.64.1",
    "peft==0.4.0",
    "transformers==4.32.0",
    "accelerate==0.21.0",
    "tokenizers==0.13.3",
    "langchain==0.0.270",
    "bitsandbytes==0.41.1",
    "numpy==1.23.5",
    "pandas==1.5.3",
    "polars==0.19.2",
    "scipy==1.11.2",
    "trl==0.5.0",
    "scikit-learn==1.3.0",

    # for OpenBook
    "sentence-transformers==2.2.2",
    "blingfire==0.1.8",

    # for deberta
    "protobuf==4.24.3",

    "plotly==5.16.1",

    "build~=0.10.0",
    "virtualenv~=20.24.3",
]
dash = [
    "dash[diskcache]~=2.12.1",
    "dash_bootstrap_components~=1.4.1",
    "dash_auth~=1.4.1"
]
train = [
    "trl==0.6.0"
]
pycharm = [
#    "git+https://github.com/yomura-yomura/plotly_utility",
    "kaleido ~= 0.2.1"
]
wiki-article = [
    "umap-learn ~= 0.5.4"
]

[tool.setuptools.packages.find]
exclude = ["data", "build", "dist", "src", "meta"]

[tool.isort]
include_trailing_comma = true
multi_line_output = 3
line_length = 120
profile = "black"

[tool.black]
line-length = 120

[tool.mypy]
show_error_context = true
strict = true
show_column_numbers = true
namespace_packages = true
disallow_subclassing_any = false
disallow_untyped_decorators = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
follow_imports = "silent"
exclude = []
